fix: 17 bug fixes across server and evr protocol packages
- Deadlock: HealthCheck calls write-lock RemoveBroadcaster while holding read-lock
- Slice corruption: filterRemoteLogs reuses caller's backing array via logs[:0]
- Inverted error check: LobbyType only assigned when StreamNumber fails
- Nil deref: ServiceSettings() not nil-checked in evr_match.go and evr_pipeline_login.go
- Error swallowed: StreamStringTable uses semicolon instead of := for error
- Error dropped: LobbyPingRequest.Stream discards RunErrorFunctions return value
- Silent marshal failure: JSON marshal error shadowed by := in session_ws.go Send()
- Auth bypass: missing return after 401 in evr_runtime_api.go
- Goroutine leak: unbuffered channels in IP info providers block on timeout
- Inverted logic: generateLinkTicket always creates duplicates (found starts true)
- Nil err in error message: discord refresh formats nil err on non-200 status
- Wrong expiry: token.Expiry.Second() returns second-of-minute, not seconds-until
- Nil deref before nil check: loginHistoryCompare dereferences before nil guard
- Nil deref: LobbyGroup.GetLeader and MatchmakerAdd missing nil guard on ph
- Nil deref: NewSignalEnvelope returns nil, callers chain .String() unchecked
- Panic: unchecked type assertion in NewPlayerParticipation
- Wrong formula: VRMLServerScore hilo operator precedence produces unnormalized values

+++ server/evr_lobby_builder_serverscore.go
@@ -183,7 +183,7 @@ func VRMLServerScore(latencies [][]float64, minRTT, maxRTT, thresholdRTT float64
 	lobbySize := float64(teamSize) * 2
 
 	// High/low ping points
-	hilo := float64((blueSum + orangeSum) - (minRTT*lobbySize)/float64((thresholdRTT*lobbySize)-(minRTT*lobbySize)))
+	hilo := ((blueSum + orangeSum) - (minRTT * lobbySize)) / ((thresholdRTT * lobbySize) - (minRTT * lobbySize))
 	hiloPoints := (1 - hilo) * pointsDistro[3]
 
 	// Final score
